+++ CHANGELOG.md
@@ -8,6 +8,12 @@ and this project adheres to
 
 ## [Unreleased]
 
+
+### Fixed
+
+- Cancel a batch order through Joanie's backoffice aborts
+  signing procedure at the signature provider
+
 ## [3.2.0] - 2026-02-25
 
 ### Added